requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
langchain
openai
psycopg2
langchain_openai